# test_output/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ilma API",
    "description": "School management system API with end-to-end encryption for messaging. Currently under development.",
    "contact": {
      "name": "API Support / Developer",
      "email": "jero.lampila@gmail.com"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/messages/threads": {
      "get": {
        "tags": [
          "messages"
        ],
        "operationId": "list_threads",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Offset for pagination",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of threads",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ThreadPreview"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "messages"
        ],
        "operationId": "send_message",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Message sent, thread created if needed"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/messages/threads/{thread_id}": {
      "get": {
        "tags": [
          "messages"
        ],
        "operationId": "get_thread_messages",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "description": "Thread UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of messages",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "before",
            "in": "query",
            "description": "Message ID to paginate before",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Messages list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/permissions": {
      "get": {
        "tags": [
          "permissions"
        ],
        "operationId": "list_permissions",
        "responses": {
          "200": {
            "description": "List of permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Permission"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/permissions/sets": {
      "get": {
        "tags": [
          "permissions"
        ],
        "operationId": "list_permission_sets",
        "responses": {
          "200": {
            "description": "List of permission sets with permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionSet"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/users/{id}/permissions": {
      "get": {
        "tags": [
          "permissions"
        ],
        "operationId": "get_user_permissions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permissions list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Permission"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "permissions"
        ],
        "operationId": "assign_user_permissions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignPermissionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Permissions assigned"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignPermissionsRequest": {
        "type": "object",
        "required": [
          "permission_ids"
        ],
        "properties": {
          "permission_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
...
      "EncryptedKey": {
        "type": "object",
        "required": [
          "recipient_id",
          "encrypted_key"
        ],
        "properties": {
          "encrypted_key": {
            "type": "string"
          },
          "recipient_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "Message": {
        "type": "object",
        "required": [
          "id",
          "thread_id",
          "sender_id",
          "sent_at",
          "ciphertext",
          "encrypted_keys"
        ],
        "properties": {
          "ciphertext": {
            "type": "string"
          },
          "encrypted_keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EncryptedKey"
            }
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sender_id": {
            "type": "string",
            "format": "uuid"
          },
          "sent_at": {
            "type": "string",
            "format": "date-time"
          },
          "thread_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "Permission": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "PermissionSet": {
        "type": "object",
        "required": [
          "id",
          "name",
          "permissions"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
...
      "SendMessageRequest": {
        "type": "object",
        "required": [
          "participant_ids",
          "ciphertext",
          "encrypted_keys"
        ],
        "properties": {
          "ciphertext": {
            "type": "string"
          },
          "encrypted_keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EncryptedKey"
            }
          },
          "participant_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "ThreadPreview": {
        "type": "object",
        "required": [
          "thread_id",
          "last_message_preview",
          "last_message_at"
        ],
        "properties": {
          "last_message_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_message_preview": {
            "type": "string"
          },
          "thread_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}